separated parseDirectoryFiles form cccsvparser to fileSystem, added function to utils.js to shorten path.resolve and path.join
